requirements.txt
requests
pyyaml
loguru
pydantic
openai
tiktoken
networkx
matplotlib
pydantic